fix: extract the request/result modifiers to different files to make code be clearer
